requirements.txt
cffi==1.15.0
msgpack==1.0.3
py-algorand-sdk==1.10.0
pycparser==2.21
pycryptodomex==3.14.1
PyNaCl==1.5.0
python-dotenv==0.19.2
tinyman-py-sdk==0.0.5
